Reorganize Lexical Menu Helpers
