+++ packages/lexical-playground/src/plugins/MentionsPlugin/index.tsx
@@ -9,8 +9,8 @@
 import {useLexicalComposerContext} from '@lexical/react/LexicalComposerContext';
 import {
   LexicalTypeaheadMenuPlugin,
-  QueryMatch,
-  TypeaheadOption,
+  MenuOption,
+  MenuTextMatch,
   useBasicTypeaheadTriggerMatch,
 } from '@lexical/react/LexicalTypeaheadMenuPlugin';
 import {TextNode} from 'lexical';
@@ -537,7 +537,7 @@ function useMentionLookupService(mentionString: string | null) {
 function checkForCapitalizedNameMentions(
   text: string,
   minMatchLength: number,
-): QueryMatch | null {
+): MenuTextMatch | null {
   const match = CapitalizedNameMentionsRegex.exec(text);
   if (match !== null) {
     // The strategy ignores leading whitespace but we need to know it's
@@ -559,7 +559,7 @@ function checkForCapitalizedNameMentions(
 function checkForAtSignMentions(
   text: string,
   minMatchLength: number,
-): QueryMatch | null {
+): MenuTextMatch | null {
   let match = AtSignMentionsRegex.exec(text);
 
   if (match === null) {
@@ -582,12 +582,12 @@ function checkForAtSignMentions(
   return null;
 }
 
-function getPossibleQueryMatch(text: string): QueryMatch | null {
+function getPossibleQueryMatch(text: string): MenuTextMatch | null {
   const match = checkForAtSignMentions(text, 1);
   return match === null ? checkForCapitalizedNameMentions(text, 3) : match;
 }
 
-class MentionTypeaheadOption extends TypeaheadOption {
+class MentionTypeaheadOption extends MenuOption {
   name: string;
   picture: JSX.Element;
