+++ packages/lexical-playground/vite.prod.config.js
@@ -135,6 +135,7 @@ const moduleResolution = [
   'LexicalCollaborationPlugin',
   'LexicalHistoryPlugin',
   'LexicalTypeaheadMenuPlugin',
+  'LexicalNodeMenuPlugin',
   'LexicalTablePlugin',
   'LexicalLinkPlugin',
   'LexicalListPlugin',
