+++ packages/lexical-react/flow/LexicalNodeMenuPlugin.js.flow
@@ -21,20 +21,12 @@ export type MenuResolution = {
   getRect: () => ClientRect,
 };
 
-export const PUNCTUATION: string =
-  '\\.,\\+\\*\\?\\$\\@\\|#{}\\(\\)\\^\\-\\[\\]\\\\/!%\'"~=<>_:;';
-
 declare export class MenuOption {
   key: string;
   ref: {current: HTMLElement | null};
   constructor(key: string): void;
   setRefElement(element: HTMLElement | null): void;
 }
-
-declare export var SCROLL_TYPEAHEAD_OPTION_INTO_VIEW_COMMAND: LexicalCommand<{
-  index: number,
-  option: MenuOption,
-}>;
 
 export type MenuRenderFn<TOption> = (
   anchorElementRef: {current: HTMLElement | null},
@@ -47,13 +39,12 @@ export type MenuRenderFn<TOption> = (
   matchingString: string,
 ) => React.Portal | React.MixedElement | null;
 
-declare export function useBasicTypeaheadTriggerMatch(
-  trigger: string,
-  options: {minLength?: number, maxLength?: number},
-): TriggerFn;
+export type TriggerFn = (
+  text: string,
+  editor: LexicalEditor,
+) => MenuTextMatch | null;
 
-export type TypeaheadMenuPluginProps<TOption> = {
-  onQueryChange: (matchingString: string | null) => void,
+type NodeMenuPluginProps<TOption> = {
   onSelectOption: (
     option: TOption,
     textNodeContainingQuery: TextNode | null,
@@ -61,18 +52,13 @@ export type TypeaheadMenuPluginProps<TOption> = {
     matchingString: string,
   ) => void,
   options: Array<TOption>,
-  menuRenderFn: MenuRenderFn<TOption>,
-  triggerFn: TriggerFn,
-  onOpen?: (resolution: MenuResolution) => void,
+  nodeKey: NodeKey | null,
   onClose?: () => void,
+  onOpen?: (resolution: MenuResolution) => void,
+  menuRenderFn: MenuRenderFn<TOption>,
   anchorClassName?: string,
 };
 
-export type TriggerFn = (
-  text: string,
-  editor: LexicalEditor,
-) => MenuTextMatch | null;
-
-declare export function LexicalTypeaheadMenuPlugin<TOption>(
-  options: TypeaheadMenuPluginProps<TOption>,
+declare export function LexicalNodeMenuPlugin<TOption>(
+  options: NodeMenuPluginProps<TOption>,
 ): React.MixedElement | null;
